requirements.txt
asgiref==3.8.1
click==8.1.7
colorama==0.4.6
cssbeautifier==1.15.1
Django==5.1
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djlint==1.34.1
EditorConfig==0.12.4
html-tag-names==0.1.2
html-void-elements==0.1.0
jsbeautifier==1.15.1
json5==0.9.25
pathspec==0.12.1
PyJWT==2.9.0
PyYAML==6.0.2
regex==2023.12.25
six==1.16.0
sqlparse==0.5.1
tqdm==4.66.5
tzdata==2024.1
